+++ projects/evo-ui-kit/src/lib/components/evo-popover/evo-popover.component.ts
@@ -1,21 +1,10 @@
-import {
-    Component,
-    HostListener,
-    Input,
-    ElementRef,
-    NgZone,
-    AfterViewInit,
-    OnChanges,
-    OnDestroy,
-    SimpleChanges,
-    ViewChild
-} from '@angular/core';
-import Popper from 'popper.js';
+import { Component, HostListener, Input, ElementRef, NgZone, AfterViewInit, OnChanges, OnDestroy, SimpleChanges, ViewChild } from '@angular/core';
+import { createPopper, Instance, Placement, Modifier } from '@popperjs/core';
+import { Subject } from 'rxjs';
+import { tap, observeOn, takeUntil } from 'rxjs/operators';
+import { async } from 'rxjs/internal/scheduler/async';
 
-const DEFAULT_DELAY = {
-    show: 0,
-    hide: 100,
-};
+export type EvoPopoverPosition = 'center' | Placement;
 
 export type EvoPopoverPosition = 'center' | Popper.Placement;
 
@@ -23,6 +12,11 @@ export interface EvoPopoverDelay {
     show?: number;
     hide?: number;
 }
+
+const DEFAULT_DELAY = {
+    show: 0,
+    hide: 100,
+};
 
 @Component({
     selector: 'evo-popover',
@@ -36,15 +30,11 @@ export class EvoPopoverComponent implements AfterViewInit, OnChanges, OnDestroy 
     @Input('media-tablet-position') mediaTabletPosition: 'right' | 'left' | 'center' = 'center';
 
     @Input('position') set position(position: EvoPopoverPosition) {
-        this.placement = position === 'center' ? (this.positionMap[position] as Popper.Placement) : position;
+        this.placement = position === 'center' ? (this.positionMap[position] as Placement) : position;
     }
 
     @Input() show = false;
-    @Input() positionFixed = false;
-    @Input() eventsEnabled = true;
-    @Input() modifiers: Popper.Modifiers;
-    @Input() target: string | Element;
-
+    @Input() modifiers: Partial<Modifier<any>>[] = [];
     @Input('delay') set setDelay(value: any) {
         if (typeof value === 'number' && value >= 0) {
             this.delay = {
@@ -69,31 +59,15 @@ export class EvoPopoverComponent implements AfterViewInit, OnChanges, OnDestroy 
     @ViewChild('popover') el: ElementRef;
     @ViewChild('popoverWrap') popoverWrap: ElementRef;
 
-    private popper: Popper;
-    private placement: Popper.Placement = 'bottom';
+    private popper: Instance;
+    private placement: Placement = 'bottom';
     private delay: EvoPopoverDelay = {};
     private visibilityTimeout = null;
-
     // Old API Map
-    private positionMap = {'center': 'bottom'};
-    private arrowSize = 16;
-    private borderRadius = 6;
-    private defaultModifiers = {
-        offset: {
-            // Modifier for Arrow offset
-            fn: data => {
-                const {offsets, placement, arrowStyles} = data;
-                const {reference, popper} = offsets;
-                if (placement === 'bottom-start' && ((reference.width / 2) + this.arrowSize) > popper.width) {
-                    arrowStyles.left = popper.width - this.arrowSize - this.borderRadius + 'px';
-                }
-                if (placement === 'bottom-end' && ((reference.width / 2) + this.arrowSize) > popper.width) {
-                    arrowStyles.left = this.borderRadius + 'px';
-                }
-                return data;
-            }
-        }
-    };
+    private positionMap = { 'center': 'bottom' };
+    private popoverVisibilityTimeout = false;
+    private update$ = new Subject();
+    private subscriptions$ = new Subject();
 
     constructor(
         private zone: NgZone,
@@ -102,20 +76,26 @@ export class EvoPopoverComponent implements AfterViewInit, OnChanges, OnDestroy 
 
     ngAfterViewInit() {
         this.create();
+        this.update$.pipe(
+            takeUntil(this.subscriptions$),
+            observeOn(async),
+            tap(() => {
+                if (this.popper) {
+                    this.popper.update();
+                }
+            })
+        ).subscribe();
     }
 
     ngOnDestroy() {
         this.destroy();
+        this.subscriptions$.next();
+        this.subscriptions$.complete();
     }
 
     ngOnChanges(changes: SimpleChanges) {
-        const {target, position, positionFixed, eventsEnabled} = changes;
-        if (
-            target && !target.firstChange ||
-            position && !position.firstChange ||
-            positionFixed && !positionFixed.firstChange ||
-            eventsEnabled && !eventsEnabled.firstChange
-        ) {
+        const position = changes.position;
+        if ( position && !position.firstChange ) {
             this.destroy();
             this.create();
         }
@@ -123,21 +103,15 @@ export class EvoPopoverComponent implements AfterViewInit, OnChanges, OnDestroy 
 
     create() {
         this.zone.runOutsideAngular(() => {
-            const {placement, positionFixed, eventsEnabled, modifiers} = this;
-
-            this.popper = new Popper(
-                this.getTargetNode(),
+            this.popper = createPopper(
+                this.el.nativeElement,
                 this.popoverWrap.nativeElement,
                 {
-                    placement,
-                    positionFixed,
-                    eventsEnabled,
-                    modifiers: {
-                        ...this.defaultModifiers,
-                        ...modifiers,
-                    },
+                    placement: this.placement,
+                    modifiers: this.modifiers,
                 }
             );
+            this.update$.next();
         });
     }
 
@@ -205,17 +179,6 @@ export class EvoPopoverComponent implements AfterViewInit, OnChanges, OnDestroy 
             }, delayValue);
         } else {
             toggle();
-        }
-    }
-
-    private getTargetNode(): Element {
-        if (!this.target) {
-            return this.el.nativeElement;
-        }
-        if (typeof this.target === 'string') {
-            return document.querySelector(this.target);
-        } else {
-            return this.target;
         }
     }