+++ projects/evo-ui-kit/src/lib/components/evo-segmented-bar/evo-segmented-bar-button/evo-segmented-bar-button.component.ts
@@ -17,7 +17,8 @@ import { EvoBaseControl } from '../../../common/evo-base-control';
 export class EvoSegmentedBarButtonComponent extends EvoBaseControl implements ControlValueAccessor, AfterViewChecked {
     @Input() name: string;
     @Input() value: string;
-    @Input() color = 'link';
+    @Input() color = 'white';
+    @Input() disabled = false;
 
     private _selectedValue: any;
 
@@ -62,6 +63,10 @@ export class EvoSegmentedBarButtonComponent extends EvoBaseControl implements Co
 
         if (this.color) {
             classes.push(this.color);
+        }
+
+        if (this.disabled) {
+            classes.push('disabled');
         }
 
         return classes;