+++ projects/evo-ui-kit/src/lib/pipes/public-api.ts
@@ -1,1 +1,1 @@
-import './evo-pipes.module';
+export * from './evo-pipes.module';