+++ src/stories/components/evo-stepper.stories.ts
@@ -3,7 +3,17 @@ import { EvoStepperModule, EvoButtonModule } from '@evo/ui-kit';
 import { Component, OnInit } from '@angular/core';
 
 
-// Component for testing template outlet
+const styles = `
+.wrap {
+    max-width: 800px;
+    margin: 40px auto 0;
+    padding: 0 40px;
+}
+.step-content, p {
+    text-align: center;
+    margin-top: 15px;
+}`;
+
 @Component({
     selector: 'test-component',
     template: `
@@ -27,8 +37,8 @@ storiesOf('Components/Stepper', module)
     )
     .add('default', () => ({
         template: `
-        <style> .step-content, pre { text-align: center; } </style>
-        <evo-stepper class="stepper" [currentStepIndex]="currentStepIndex" (onChange)="onStepChanged($event)">
+        <style>${ styles }</style>
+        <evo-stepper class="stepper wrap" [currentStepIndex]="currentStepIndex" (onChange)="onStepChanged($event)">
             <evo-stepper-item label="One">
                 <div class="step-content">
                     <h3 class="evo-title evo-title_h3">Step 1</h3>
@@ -79,7 +89,7 @@ storiesOf('Components/Stepper', module)
             </evo-stepper-item>
         </evo-stepper>
         <br>
-        <pre>Get step index from output: {{stepNumber}}</pre>
+        <pre class="wrap">Get step index from output: {{stepNumber}}</pre>
         `,
         props: {
             steps: [ '', 'Two', 'Three', 'Four', 'Five', 'Finish' ],
@@ -93,8 +103,8 @@ storiesOf('Components/Stepper', module)
     .add('with life cycle hooks on step changes', () => ({
         /* tslint:disable */
         template: `
-        <style> .step-content, p { text-align: center; margin-top: 15px; } button { display: block; margin: 15px auto; }</style>
-        <evo-stepper class="stepper" [currentStepIndex]="currentStepIndex">
+        <style>${ styles }  button { display: block; margin: 15px auto; }</style>
+        <evo-stepper class="stepper wrap" [currentStepIndex]="currentStepIndex">
             <evo-stepper-item label="One">
                 <div class="step-content">
                     <h3 class="evo-title evo-title_h3">Step 1</h3>
@@ -116,6 +126,100 @@ storiesOf('Components/Stepper', module)
         /* tslint:enable */
         props: {
             currentStepIndex: 0,
+        },
+    }))
+    .add('with loop', () => ({
+        /* tslint:disable */
+        template: `
+        <style>${ styles }</style>
+        <evo-stepper class="wrap"
+            [currentStepIndex]="currentStepIndex"
+            [clickableItems]="clickableItems">
+            <evo-stepper-item *ngFor="let step of steps; index as i" [label]="step.label">
+                <div class="step-content">
+                    <h3 class="evo-title evo-title_h3">{{ step.title }}</h3>
+                    <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Repellendus corrupti quasi recusandae culpa nisi iusto, ab rem illo magni maxime mollitia eveniet quam magnam doloribus vel corporis? Perferendis, officia sunt.</p>
+                </div>
+            </evo-stepper-item>
+        </evo-stepper>
+        <div class="wrap">
+            <button evo-button (click)="next()" size="small">Next</button>
+            &nbsp;&nbsp;
+            <button evo-button [disabled]="!newSteps.length" (click)="addStep()" size="small">Add Step</button>
+        </div>
+        `,
+        /* tslint:enable */
+        props: {
+            currentStepIndex: 0,
+            steps: [{
+                label: 'One',
+                title: 'First text',
+            }, {
+                label: 'Two',
+                title: 'Second text',
+            }],
+            newSteps: [{
+                label: 'Three',
+                title: 'Third text',
+            }, {
+                label: 'Four',
+                title: 'Some another text',
+            }],
+            next: function() {
+                if ((this.steps.length - 1) > this.currentStepIndex) {
+                    this.currentStepIndex += 1;
+                }
+            },
+            addStep: function() {
+                if (this.newSteps.length) {
+                    this.steps.push(this.newSteps.splice(0, 1)[0]);
+                }
+            }
+        },
+    }))
+    .add('with clickable step indicators', () => ({
+        /* tslint:disable */
+        template: `
+        <style>${ styles }</style>
+        <evo-stepper class="wrap"
+            [currentStepIndex]="currentStepIndex"
+            [clickableItems]="clickableItems"
+            (clickItem)="handleClick($event)">
+            <evo-stepper-item *ngFor="let step of steps; index as i" [label]="step.label">
+                <div class="step-content">
+                    <h3 class="evo-title evo-title_h3">{{ step.title }}</h3>
+                    <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Repellendus corrupti quasi recusandae culpa nisi iusto, ab rem illo magni maxime mollitia eveniet quam magnam doloribus vel corporis? Perferendis, officia sunt.</p>
+                </div>
+            </evo-stepper-item>
+        </evo-stepper>
+        <div class="wrap">
+            <button evo-button (click)="next()" size="small">Next</button>
+        </div>
+        `,
+        /* tslint:enable */
+        props: {
+            currentStepIndex: 0,
+            clickableItems: true,
+            steps: [{
+                label: 'One',
+                title: 'First text',
+            }, {
+                label: 'Two',
+                title: 'Second text',
+            }, {
+                label: 'Three',
+                title: 'Third text',
+            }],
+            handleClick: function(index: number) {
+                if (this.currentStepIndex > index) {
+                    this.currentStepIndex = index;
+                }
+            },
+            next: function() {
+                if ((this.steps.length - 1) > this.currentStepIndex) {
+                    this.currentStepIndex += 1;
+                }
+            },
         },
     }));